requirements.txt
torch
kafka-python
sentence-transformers
confluent-kafka
python-dotenv
coverage
textsum
testcontainers